Add force parameter to AbstractContext.set_path
- Can be used if proper path has already been created
