+++ resources/lib/youtube_plugin/kodion/utils/__init__.py
@@ -12,7 +12,6 @@ from __future__ import absolute_import, division, unicode_literals
 
 from . import datetime_parser
 from .methods import (
-    create_path,
     duration_to_seconds,
     find_best_fit,
     find_video_id,
@@ -35,7 +34,6 @@ from .system_version import current_system_version
 
 
 __all__ = (
-    'create_path',
     'current_system_version',
     'datetime_parser',
     'duration_to_seconds',
