+++ resources/lib/youtube_plugin/kodion/utils/methods.py
@@ -18,12 +18,11 @@ import shutil
 from datetime import timedelta
 from math import floor, log
 
-from ..compatibility import byte_string_type, quote, string_type, xbmc, xbmcvfs
+from ..compatibility import byte_string_type, string_type, xbmc, xbmcvfs
 from ..logger import log_error
 
 
 __all__ = (
-    'create_path',
     'duration_to_seconds',
     'find_best_fit',
     'find_video_id',
@@ -174,24 +173,6 @@ def select_stream(context,
         context.log_debug('selected stream: %s' % log_data)
 
     return selected_stream_data
-
-
-def create_path(*args, **kwargs):
-    path = '/'.join([
-        part
-        for part in [
-            str(arg).strip('/').replace('\\', '/').replace('//', '/')
-            for arg in args
-        ] if part
-    ])
-    if path:
-        path = path.join(('/', '/'))
-    else:
-        return '/'
-
-    if kwargs.get('is_uri', False):
-        return quote(path)
-    return path
 
 
 def strip_html_from_text(text):
